third_party/WebKit/LayoutTests/compositing/always-composite-fixed-position-when-descendants-composite.html

<!doctype HTML>
Any errors will show below this line.
<!-- Having a composited child should always be a position:fixed compositing trigger -->
<div style="position: fixed; width: 100px; height: 100px; background: lightgray">
    <div style="will-change: transform; margin: 50px; width: 50px; height: 50px; background: lightblue">
    </div>
</div>
<script src="../resources/testharness.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals) {
    var layers = JSON.parse(internals.layerTreeAsText(document));
    assert_true(layers["layers"].length == 3);
}
</script>
